feat(pod5): read POD5 V6, keep writing V5 until the ecosystem catches up
Upstream 0.3.46 introduced POD5 V6, whose sole change is retyping the
reads-table `channel` column from uint16 to uint32 — same name, same
position. Because it retypes an existing column instead of appending new
ones the way V4 and V5 did, it is not a change a narrow reader degrades
past: pinned to uint16, every V6 file fails outright.

Read side: `channel` resolves to whichever width the file carries and
widens to u32 on all three paths — per-row `ReadsBatchView::read`, the
bulk `append_columns`, and `BatchFieldExtractor` (the `read_iter` route).
`ReadData.channel`, `ReadColumns.channel`, and the Python
`ReadData.channel` / `Writer.add_read` parameter become u32, matching
upstream's own C++ `ReadData`; numpy exports move uint16 -> uint32.

Write side stays V5 deliberately. The newest `pod5` on PyPI is 0.3.44 and
rejects a uint32 channel outright ("Schema field 'channel' is incorrect
type: 'uint32'") — verified by writing a V6 file and opening it with the
env's reference reader. Emitting V6 today would make every escpod-written
file unreadable by every installable reader, to carry channel numbers no
flow cell produces (PromethION tops out at 3000). So the emitted column
stays uint16, files stay stamped 0.3.44, and the one input that would
actually lose data — a channel above u16::MAX, reachable only from a
genuine V6 file — is refused by `schema::reads::narrow_channel` rather
than written as `channel % 65536`. That is the single site to flip when
v6-capable wheels ship, and a test pins the schema width to POD5_VERSION
so the two cannot drift apart.

`build_partition_inner` (merge/filter/subset) becomes fallible to carry
that error; `escpod inspect summary` channel stats widen to u32.

Tests: there is no V6 fixture to be had — upstream's test_data stops at
v5 and no released tool can write one — so `reader::v6_compat` builds a
genuine one, transcoding an escpod-written V5 file's reads table to
uint32 (with channels up to 4e9) and re-assembling the container, then
opens it through the ordinary `Reader`. Both V6 guardrails were confirmed
to fail against the uint16-pinned reader before the fix. Plus per-width
column tests, a writer refusal test, and a PromethION-range round-trip.

Verified: clippy/fmt clean; 690 workspace tests + 23 doctests pass; the
POD5 reference-reader compat suite and the Python binding suite pass.

+++ crates/escapepod-cli/src/commands/summary.rs
@@ -32,8 +32,8 @@ struct ReadStats {
     count: u64,
     total_samples: u64,
     lengths: Vec<u64>,
-    channels: HashMap<u16, u64>,
-    max_channel: u16,
+    channels: HashMap<u32, u64>,
+    max_channel: u32,
     wells: [u64; 5], // Index 0 unused, wells 1-4
     end_reasons: HashMap<String, u64>,
 }
@@ -87,7 +87,7 @@ struct StatisticsSummary {
     length_median: u64,
     length_n50: u64,
     active_channels: usize,
-    total_channels: u16,
+    total_channels: u32,
 }
 
 /// Current POD5 version for comparison.
@@ -297,7 +297,7 @@ pub fn run(args: SummaryArgs) -> anyhow::Result<()> {
 /// Known flowcell sizes:
 /// - MinION/Flongle: 512 channels
 /// - PromethION/P2: 2675 or 3000 channels
-fn infer_total_channels(max_channel: u16) -> u16 {
+fn infer_total_channels(max_channel: u32) -> u32 {
     if max_channel <= 512 {
         512
     } else if max_channel <= 2675 {
